Add files via upload
Motor Club of America (MCA) members are protected no matter what might happen to them on the road. With true unlimited roadside assistance, members can quickly get 24/7 help for car trouble or an accident and unlimited free towing to the nearest service facility. Should they receive a ticket, we are the only service with attorneys who will fight that ticket in court. Stranded far from home as the result of an accident? We’ll pay up to $500 for food, lodging or transportation.

MCA members are covered on Emergency Room visits and in-hospital stays resulting from a covered accident or, should the worst happen, a $10,000 accident death benefit.